pyproject.toml
[tool.poetry]
name = "shopping-data"
version = "0.1.0"
description = ""
authors = ["Pedro Rocha <pedro.diniz.rocha@gmail.com>"]
readme = "README.md"
packages = [{include = "shopping_data"}]

[tool.poetry.dependencies]
python = "^3.11"
pendulum = "^2.1.2"
mypy = "^1.0.1"
sqlalchemy = "^2.0.17"
alembic = "^1.11.1"
pymysql = "^1.0.3"
python-dotenv = "^1.0.0"
ruff = "^0.0.275"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
